Use correct dtos in CommentsController. Change in CommentService implementation.
The CommentsControllers PutComment method now uses an UpdateCommentDTO instead of the Comment model itself. The PostComment returns a CommentDTO instead of a Comment. The interface and implementation of the CommentServices update methods signature has been changed to only use a comment as a paramater, since the id parameter is not used